Add --caldav-passwd-cmd option
I don't use password-store and don't like having my passwords in process
listings (which easily happens if using the $CALDAV_PASSWD env var).

This patch offers an alternative, by specifying a command that outputs
the password on stdout, for example:

  --caldav-passwd-cmd="cat /run/secrets/caldav-pw"

The command is run through a shell, so "cat ~/mypassword" also works.