Change CalDAV description format
Old mappings:
- TW `annotations`, `uuid` <-> Caldav `DESCRIPTION`

When a TW item is converted to caldav item, the `annotations` and `uuid`
of TW item are encoded into the following format, and stored in
`DESCRIPTION` of caldav.
```plaintext
IMPORTED FROM TASKWARRIOR

* Annotation 1: first annotation
* Annotation 2: second annotation
* Annotation 3: third annotation
* uuid: 12345678-123-1234-1234-1234567890ab
```

New mappings:
- TW `uuid` <-> Caldav `X-SYNCALL-TW-UUID`
- TW `annotations` <-> Caldav `DESCRIPTION`
    (each annotation <-> a line in description)

In this new mappings, the `uuid` of TW item will be mapped to a
non-standard iCalendar field `X-SYNCALL-TW-UUID`, allowed by RFC 5545.
The `annotation` of TW item will be mapped to `DESCRIPTION` of caldav
item. Each line in `DESCRIPTION` is corresponding to an annotation.